Fix authorization problem in topic tree finder for sub sites.
The schema extender is always anonymous.
Using listFolderContents takes authorization in account, resulting in that not published
topic trees are never visible for anyone and therefore the subsite topics are not selectable
when there are no trees in the subsite yet published.

Therefore the check for trees in subsite should not be security dependent, using objectValues now.