fix dexterity behavior and relation collector.
When using the behavior on a Dexterity object and thus having DX to DX
relations, the collector was unable to lookup the backreferences since
it did only use the reference_catalog, which is an Archetypes tool and
does not track DX to DX references.

This change installes plone.app.relationfield with zc.relation and uses
the zc.relation catalog (intid based) for looking up backreferences.